feat: changed prefix filter to search filter, allow all/invert with ctrl
Instead of a prefix filter, the search filter is now searched within all entries. This seems to be more common than a prefix search.

Using the search functionality disabled the ability to select all options or invert the selection in the checkbox control. If the search filter is enabled, these two functionalities can now be used with the key modifier ctrl.

Updated the displayed instructions to match the changes made.