fix: allow `None` as argument for `kbi_msg`
This commit addresses the issue of the keyboard interrupt message always printing,
even in cases where it's not needed and is explicitly set to an empty string.
This allows users to disable/replace said print when needed, while avoiding the redundant
newline that's currenly being printed when `kbi_msg=""`.

This commit also introduces the usage of `TYPE_CHECKING` to improve performance
in the changed modules.